% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot.gp_model}
\alias{plot.gp_model}
\title{Plot fitted GP model}
\usage{
\method{plot}{gp_model}(
  model,
  level = c("ind", "group", "global"),
  .width = c(0.5, 0.8, 0.99),
  n_samples = 0,
  prior = FALSE,
  show_ci = TRUE,
  facet = show_ci,
  show_events = TRUE,
  dev_only = FALSE,
  show_traces = TRUE,
  palette = "Purples",
  width = 0.1,
  height = 0.2,
  size = 10,
  sample_alpha = 0.6,
  sample_colour = "darkgrey",
  sample_linewidth = 0.9,
  rescale = FALSE,
  plot_mu = TRUE
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{level}{Character. The level (\code{"ind"}, \code{"group"}, \code{"global"}) at which
the posterior fitted Gaussian Processes are extracted. By default, the level is \code{"ind"},
thereby showing the additive combination of group-level and individual GPs translated by
their respective baselines, \code{mu_ind[i]}, where i is the individual ID.}

\item{.width}{Numeric vector. Highest density posterior credible interval bands around
the mean GP trace to visualise. Strictly, widths show pointwise quantiles around
the mean. By default, \code{.width = c(0.5, 0.8, 0.99)}, marking 50\%, 80\% and 99\% density
coverage (see \code{\link[ggdist:interval_widths]{ggdist::interval_widths()}}).}

\item{n_samples}{Integer. Number of samples to draw to compute prior/posterior
predictive checks. By default this is 1000 and can be no greater than the number of
post-warmup iterations times the number of chains (see \code{\link[=gp_fit]{gp_fit()}}). Using the default
parameters means this ceiling is 4000 samples.}

\item{prior}{Boolean. Determines whether to show GPs drawn from prior or posterior
parameters. By default, \code{prior=FALSE} (see \code{\link[=gp_fit]{gp_fit()}}).}

\item{facet}{Boolean. Controls whether plots are overlaid onto the same axes or
faceted by the level specified. By default, this is \code{TRUE} but is set to \code{FALSE}
if \code{show_cri} is \code{FALSE}. This behaviour can be overridden by setting \code{facet=TRUE}.}

\item{show_events}{Boolean. Determines whether event data, if available are shown
as subplots beneath each plot of GP traces. By default, if event data are available,
they are plotted (\code{show_events=TRUE}).}

\item{dev_only}{Boolean. Determines whether only the deviations relative to
the level above are to be plotted. For example, if \code{dev_only=TRUE} and \code{level="ind"}
in a single-group model, only \eqn{\mu_{ind[i]} + f_{ind[i]}} would be shown
for the ith individual, rather than the composite \eqn{\mu_{ind[i]} + f_{ind[i]} + f_{group[i]}}.}

\item{show_traces}{Boolean. Determines whether simulate ground truth traces are
to be plotted, if available.}

\item{palette}{Character. Colour scheme for plotting credible intervals. The choice
of palette is passed to \code{\link[ggplot2:scale_fill_brewer]{ggplot2::scale_fill_brewer()}} (see \code{\link[RColorBrewer:display.brewer.all]{RColorBrewer::display.brewer.all()}} for
the full set of colour palettes).}

\item{width}{Numeric. Width of the event rasters. The default is 0.1. Passed to \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}}

\item{height}{Numeric. Height of the event rasters. The default is 0.2. Passed to \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}}

\item{size}{Numeric. Axis label size. Default is 10}

\item{sample_alpha}{Numeric. The alpha level for random MCMC samples drawn from the
prior or posterior model parameters. The default alpha is \code{0.6}}

\item{sample_colour}{Character. The line colour of the random MCMC samples drawn from the
prior or posterior model parameters. The default colour is \code{"darkgrey"}}

\item{sample_linewidth}{Numeric. The linewidth for random MCMC samples drawn from the
prior or posterior model parameters. The default linewidth is \code{0.9}.}

\item{rescale}{Boolean. The observed traces represent the time-modulated scale, \eqn{\theta(t)},
of the renewal process. However, for non-memoryless processes, namely, the exponential
the expected inter-event interval is also specified by shape parameters. \code{rescale}
applies a scaling constant, \eqn{c}, to the fitted GP traces (and simulated ground truth traces if applicable)
that corresponds with the survival family.
\deqn{\mathbb{E}[X] = \exp{\beta(t)} c \text{, where } \beta(t) = \ln{\theta(t)}}
By default, the traces are shown on the response scale and hence the inverse ln
is applied to \eqn{c}:
\deqn{\mathbb{E}[X] = \beta(t) + \ln{c}}
Where for the gamma family, \eqn{c = k}.
For Weibull, \eqn{c = \Gamma(1 + \frac{1}{\text{shape}})}.
For the generalised gamma, \eqn{c = \frac{\frac{\Gamma(k + 1)}{\Gamma(shape)}}{\frac{\Gamma(k)}{\Gamma(shape)}}}
Rescaling ensures that the expected inter-event interval (accounting for scale and shape)
is shown. For \code{\link[=get_nlpd]{get_nlpd()}}, rescale is by default \code{TRUE}.}

\item{show_cri}{Boolean. Controls whether credible intervals are plotted or not.
If \code{FALSE}, by default, \code{facet} will also be set to \code{FALSE}, overlaying all plots
on the same axes. By default, however, show_cri is \code{TRUE}.}
}
\description{
Plot fitted GP model
}
