% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{ppc_get_ecdf}
\alias{ppc_get_ecdf}
\title{Plot empirical cumulative density functions from prior or posterior predictive distributions to compare against the observed event data}
\usage{
ppc_get_ecdf(
  model = NULL,
  ppc_events = NULL,
  n_samples = 1000,
  plot_density = TRUE,
  spaghetti = 0,
  palette = "Purples",
  .width = c(0.5, 0.8, 0.99),
  bw_kde = 1.2,
  return_plot = TRUE,
  prior = FALSE,
  linewidth = 1,
  color = "blue",
  resolution = 0.1
)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{ppc_events}{tibble. Optional tibble of pre-drawn prior/posterior predictive
events generated via \code{\link[=ppc_draw_events]{ppc_draw_events()}}. If ppc_events is \code{NULL}, the default,
\code{model} cannot be \code{NULL} and events will be generated via the function on the passed model.}

\item{n_samples}{Integer. Number of samples to draw to compute prior/posterior
predictive checks. By default this is 1000 and can be no greater than the number of
post-warmup iterations times the number of chains (see \code{\link[=gp_fit]{gp_fit()}}). Using the default
parameters means this ceiling is 4000 samples.}

\item{palette}{Character. Colour scheme for plotting credible intervals. The choice
of palette is passed to \code{\link[ggplot2:scale_fill_brewer]{ggplot2::scale_fill_brewer()}} (see \code{\link[RColorBrewer:display.brewer.all]{RColorBrewer::display.brewer.all()}} for
the full set of colour palettes).}

\item{.width}{Numeric vector. Highest density posterior credible interval bands around
the mean GP trace to visualise. Strictly, widths show pointwise quantiles around
the mean. By default, \code{.width = c(0.5, 0.8, 0.99)}, marking 50\%, 80\% and 99\% density
coverage (see \code{\link[ggdist:interval_widths]{ggdist::interval_widths()}}).}

\item{bw_kde}{Numeric or character. The smoothing bandwidth to use for the interevent interval
distribution. The default is \code{"nrd0"} (see \code{\link[stats:density]{stats::density()}})}

\item{return_plot}{Boolean. Specify whether to return a ggplot or a list of data.
By default, this is TRUE}

\item{prior}{Boolean. Plot prior estimates, if available, instead of posterior estimates.
The default is \code{FALSE}}

\item{resolution}{Numeric. The resolution of the time grid for computing instantaneous
event rate (see \code{\link[=ppc_get_inst_eventrate]{ppc_get_inst_eventrate()}}). By default, this is 0.01}
}
\value{
A ggplot or tibble of data
}
\description{
Plot empirical cumulative density functions from prior or posterior predictive distributions to compare against the observed event data
}
\seealso{
\code{\link[=ppc_get_interevent_dist]{ppc_get_interevent_dist()}}, \code{\link[=ppc_get_eventrate]{ppc_get_eventrate()}}, \code{\link[=ppc_plot_all]{ppc_plot_all()}}.
}
