% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gps.R
\name{simulate_gp_traces}
\alias{simulate_gp_traces}
\title{Simulate Gaussian Process Trajectories}
\usage{
simulate_gp_traces(
  n_ind = 5,
  n_group = 1,
  duration = 120,
  alpha = 0.8,
  mu_ind = 0,
  mu_group = 0,
  mu_global = 0,
  sigma_ind = 0.3,
  sigma_group = 0.05,
  alpha_ind = 0.3,
  alpha_group = 5/4 * alpha_ind,
  alpha_global = 5/4 * alpha_group,
  rho_ind = 10,
  rho_group = 5/4 * rho_ind,
  rho_global = 5/4 * rho_group,
  kernel = "squared_exp",
  L_factor = 1.2,
  M,
  gp_seed = NULL,
  mu_seed = NULL
)
}
\arguments{
\item{n_ind}{Integer or integer vector. Number of individuals per group. If 1 integer is supplied,}

\item{duration}{Integer. The length of the full time series to simulate (in seconds). The default is 100 seconds}

\item{alpha_global, alpha_group, alpha_ind}{Numeric. Amplitude parameters of the
covariance kernel. \code{alpha_global}, \code{alpha_group} and \code{alpha_ind} respectively
set the amplitude terms at global, group and individual-level trajectories.
By default, alpha terms at each hierarchy are set to be 60\% of the preceding level,
rounded to the nearest tenth, starting from a maximum alpha of 0.8.
If the simulation is multi-group, \code{alpha_global = 0.8},\code{alpha_group = 0.5} and
\code{alpha_ind = 0.3}. If single-group and multi-individual, \code{alpha_group = 0.8} and
\code{alpha_ind = 0.5}. If only one alpha term is specified, the others are extrapolated
in the same geometric series. Setting \code{alpha_ind = 0.8}, for example, would set
\code{alpha_group = 1.3}.}

\item{rho_global, rho_group, rho_ind}{Numeric. Length-scale parameters of the covariance
kernel. \code{rho_global}, \code{rho_group} and \code{rho_ind} respectively
set the length-scale terms at global, group and individual-level trajectories. By default,
length-scale parameters are set at the same scale at 15. If only one rho term is specified,
the others are set to the same value.}

\item{kernel}{Character. Choice of covariance kernel, K, to simulate Gaussian Process
trajectories, comprising "squared_exp", "matern", "periodic".
The squared exponential (or radial basis function) is the default kernel.
If "matern" is specified, the default Mátern smoothness parameter, nu, is 3/2.
"matern12", "matern32" and "matern52" may also be specified, setting smoothness to
1/2, 3/2 and 5/2, respectively. Note that currently only smoothness terms at these intervals
are supported in the package. Squared exponential, Mátern and periodic kernels are
parameterised by a combination of length-scale, \eqn{\rho} (or \eqn{\ell}) and amplitude-scale,
\eqn{\alpha}.}

\item{L_factor}{Numeric.}

\item{M}{Numeric. Number of basis functions to use for Hilbert space approximation.
If unspecified, M functions are selected according to recommendations in Riutort et al. (2017)}

\item{nu}{Numeric. Mátern smoothness parameter. Typical smoothness terms are 1/2, 3/2 and 5/2
and these are the only intervals that can currently be modelled by the package.
The default is 3/2}

\item{P}{Numeric. Period duration parameter for periodic kernels.}
}
\value{
An object of class \code{sim_traces}, containing a list of \code{ind_traces},
\code{group_traces} (trace common to individuals in the same group),
\code{global_traces} (trace common to all individuals; NULL if simulation has only
1 individual or 1 group) and \code{sim_parameters} (simulation parameters).
}
\description{
Simulate Gaussian Process Trajectories
}
\examples{
gp_traces<- simulate_gp_traces(n_groups = 2, n_ind = 5, kernel = "matern32")
plot(gp_traces)

}
\seealso{
The methods \code{\link[=simulate_spline_traces]{simulate_spline_traces()}} and
\code{\link[=simulate_boxcar_traces]{simulate_boxcar_traces()}}
}
