% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{ppc_draw_events}
\alias{ppc_draw_events}
\title{Generate events from prior or posterior predictive distribution of a GP model}
\usage{
ppc_draw_events(model, n_samples, prior = FALSE, resolution = 0.1)
}
\arguments{
\item{model}{gp_model. A fitted GP model or model with prior checks}

\item{n_samples}{Integer. Number of samples to draw to compute prior/posterior
predictive checks. By default this is 1000 and can be no greater than the number of
post-warmup iterations times the number of chains (see \code{\link[=gp_fit]{gp_fit()}}). Using the default
parameters means this ceiling is 4000 samples.}

\item{prior}{Boolean. Determines whether to show GPs drawn from prior or posterior
parameters. By default, \code{prior=FALSE} (see \code{\link[=gp_fit]{gp_fit()}}).}
}
\value{
Generated events at the individual level given the prior or posterior
model parameters.
}
\description{
Taking n random MCMC draws from the prior or posterior, events are generated,
using the baseline, GP and survival parameters of the draw at the individual level
across the full duration of the time domain.
}
\seealso{
\code{\link[=draw_traces]{draw_traces()}}, \code{\link[=ppc_get_eventrate]{ppc_get_eventrate()}}, \code{\link[=ppc_get_interevent_dist]{ppc_get_interevent_dist()}},
[ppc_get_inst_eventrate().
}
